% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daDF}
\alias{daDF}
\title{Data set for example at Readme}
\format{A data frame with 200 rows and 1 variable}
\usage{
daDF
}
\description{
A dataset containing values for the example. Need to have a better description
look at example at ggplot2 at Github https://raw.githubusercontent.com/tidyverse/ggplot2/master/R/data.R
}
\keyword{datasets}
