\name{WeibullR-package}
\alias{WeibullR-package}

\docType{package}
\title{
    Weibull-R : Weibull Analysis on R
}
\description{
    Life data analysis in the graphical tradition of Waloddi Weibull
}
\author{
Jurgen Symynck <jusy@openreliability.org>
David J. Silkworth <djsilk@openreliability.org>
}

\references{
    \itemize{
        \item
            Robert B. Abernethy,
            \emph{The New Weibull Handbook, Fifth Edition}
            (Robert B. Abernethy, 2004)
        \item
            Jerald F. Lawless,
            \emph{Statistical Models and Methods for Lifetime Data,
                2nd edition}
            (Wiley-Interscience, Hoboken  N.J., 2003)
        \item William Q. Meeker and Luis A. Escobar,
            \emph{Statistical Methods for Reliability Data}
            (Wiley-Interscience, New York, 1998)
        
        \item Chi-Chao Lui,
            \emph{A Comparison Between The Weibull And Lognormal Models Used
                To Analyse Reliability Data}
            (dissertation from University of Nottingham, 1997)
%        \item
%            Efron, B. and Tibshirani, R.,
%            \emph{An Introduction to the Bootstrap}
%            (Chapman & Hall, 1993)
        \item Jurgen Symynck, Filip De Bal,
            \emph{Weibull analysis using R, in a nutshell}
            (New Technologies and Products in Machine Manufacturing Technology,
            Stefan cel Mare University of Suceava, 2010).

%            Available on
%\url{http://mechanics.kahosl.be/fatimat/index.php/downloads-and-information/40/171}.
        \item Jurgen Symynck, Filip De Bal,
        \emph{Monte Carlo pivotal confidence bounds for Weibull analysis,
            with implementations in R} (New Technologies and Products in
            Machine Manufacturing Technology, Stefan cel Mare University of
            Suceava, 2011).

            Available on
\url{http://www.fim.usv.ro/conf_1/tehnomusjournal/pagini/journal2011/files/7.pdf}
%and \url{http://mechanics.kahosl.be/fatimat/index.php/downloads-and-information/40/198}.
        
        \item \url{http://sourceforge.net/projects/weibulltoolkit/}
    }
}

\keyword{ package }
